refactor: Move ttcalc main() body into cmd/samoyed-ttcalc
ttcalc_main.go only ever existed to be invoked as the standalone
ttcalc demo binary, so move its logic directly into
cmd/samoyed-ttcalc/main.go rather than routing through an exported
TTCalcMain() function in the direwolf package.

All AX25*/AGWPEHeader/ALevel symbols it relies on were already
exported, so no wrapper file is needed - the unexported calculator/
do_lastop/connect_to_server helpers move wholesale into cmd/ instead.
