+++ cmd/samoyed-ttcalc/main_test.go
@@ -1,4 +1,4 @@
-package direwolf
+package main
 
 import (
 	"testing"
